Change type of throttleScrollCallbackMS from double to NSUInteger. onScroll should work now

+++ React/Views/RCTScrollViewManager.m
@@ -40,7 +40,7 @@ RCT_EXPORT_VIEW_PROPERTY(scrollsToTop, BOOL)
 RCT_EXPORT_VIEW_PROPERTY(showsHorizontalScrollIndicator, BOOL)
 RCT_EXPORT_VIEW_PROPERTY(showsVerticalScrollIndicator, BOOL)
 RCT_EXPORT_VIEW_PROPERTY(stickyHeaderIndices, NSNumberArray);
-RCT_EXPORT_VIEW_PROPERTY(throttleScrollCallbackMS, double);
+RCT_EXPORT_VIEW_PROPERTY(throttleScrollCallbackMS, NSUInteger);
 RCT_EXPORT_VIEW_PROPERTY(zoomScale, CGFloat);
 RCT_EXPORT_VIEW_PROPERTY(contentInset, UIEdgeInsets);
 RCT_EXPORT_VIEW_PROPERTY(scrollIndicatorInsets, UIEdgeInsets);
